+++ src/issue/Tags.ts
@@ -1,18 +1,18 @@
 export class Tags {
   private tagsMap: Map<string, Map<string, number>>;
 
-  public constructor(private readonly rawTags: { key: string; value: string }[]) {
-    this.tagsMap = new Map(this.rawTags.map(t => [t.key, new Map()]));
-
-    this.rawTags.forEach(t => {
-      const correspondingMap = this.tagsMap.get(t.key);
-      const valueOfCorrespondingMap = correspondingMap?.get(t.value);
-      if (valueOfCorrespondingMap !== undefined) {
-        correspondingMap?.set(t.value, valueOfCorrespondingMap + 1);
-      } else {
-        correspondingMap?.set(t.value, 1);
-      }
-    });
+  public constructor(
+    private readonly rawTags: {
+      key: string;
+      topValues: { count: number; value: string }[];
+    }[],
+  ) {
+    this.tagsMap = new Map(
+      this.rawTags.map(t => [
+        t.key,
+        new Map(t.topValues.map(({ count, value }) => [value, count])),
+      ]),
+    );
   }
 
   public get values(): {